requirements.txt
requests==2.32.3
pytest==8.3.4
pytest-html==4.1.1
allure-pytest==2.13.5
PyYAML==6.0.2
PyMySQL==1.1.1
jsonpath-ng==1.7.0
faker==33.3.1
loguru==0.7.3
